reformat package and split retrieve functionality to new method
