+++ simple_ars/core.py
@@ -1,10 +1,10 @@
 import logging
 
-from simpleARS import core_utils, search_object, extraction
+from simple_ars import core_utils, search_object, extraction
 
 __authors__ = 'Manolis Tsoukalas'
 __date__ = '2017-1-3'
-__version__ = '0.4'
+__version__ = '0.5'
 
 """
 simpleARS core functionality for retrieving data
@@ -74,9 +74,27 @@ def retrieve_list_data(list_data, search):
     return items
 
 
+def ars(api_response, search):
+    """
+    basic method for searching and retrieving data in dictionary format.
+    :param api_response: data returned by your endpoint
+    :param search: search object in dictionary format
+    :return: the extracted data
+    """
+    search_obj = search_object.SearchObject(search)
+
+    if isinstance(api_response, dict):
+        retrieved_data = retrieve_sub_data(api_response, search_obj)
+    elif isinstance(api_response, list):
+        retrieved_data = retrieve_list_data(api_response, search_obj)
+    else:
+        raise TypeError("Wrong Type Response!!! Response must be list or dict not {}".format(type(api_response)))
+    return retrieved_data
+
+
 def retrieve_data(api_url, search, credentials=None, mode="return", csv_file_name="csv_output"):
     """
-    entry method for retrieving data.
+    facade method for retrieving data.
     takes attributes a url a search json and mode for extraction
     :param api_url: url for the endpoint which you want to retrieve data
     :param search: search object in dictionary format
@@ -89,14 +107,8 @@ def retrieve_data(api_url, search, credentials=None, mode="return", csv_file_nam
     logging.basicConfig(level=logging.INFO)
     logger = logging.getLogger(__name__)
     api_response = core_utils.load_api_response(api_url, credentials)
-    search_obj = search_object.SearchObject(search)
 
-    if isinstance(api_response, dict):
-        retrieved_data = retrieve_sub_data(api_response, search_obj)
-    elif isinstance(api_response, list):
-        retrieved_data = retrieve_list_data(api_response, search_obj)
-    else:
-        raise TypeError("Wrong Type Response!!! Response must be list or dict not {}".format(type(api_response)))
+    retrieved_data = ars(api_response, search)
 
     if mode == "csv":
         logger.info("CSV Extraction")
